+++ R/backend_future.R
@@ -177,13 +177,15 @@ future_build <- function(
   do_prework(config = config, verbose_packages = FALSE)
   build <- try_build(target = target, meta = meta, config = config)
   if (identical(caching, "master")) {
-    build$checksum <- get_outfile_checksum(target, config)
+    build$checksum <- get_outfile_checksum(target, build$value, config)
     build <- classify_build(build, config)
     build <- serialize_build(build)
     return(build)
   }
   conclude_build(build = build, config = config)
-  list(target = target, checksum = get_checksum(target, config))
+  checksum <- get_checksum(target, build$value, config)
+  value <- hpc_worker_build_value(target, build$value, config)
+  list(target = target, value = value, checksum = checksum)
 }
 
 running_targets <- function(config) {
@@ -223,6 +225,7 @@ conclude_worker <- function(worker, config) {
   if (identical(caching, "worker")) {
     wait_checksum(
       target = build$target,
+      value = build$value,
       checksum = build$checksum,
       config = config
     )
@@ -232,6 +235,7 @@ conclude_worker <- function(worker, config) {
   }
   wait_outfile_checksum(
     target = build$target,
+    value = build$value,
     checksum = build$checksum,
     config = config
   )