+++ R/drake_meta_.R
@@ -1,8 +1,14 @@
-drake_meta_ <- function(target, config) {
+drake_meta_ <- function(target, config, spec = NULL) {
   class(target) <- drake_meta_class(target, config)
-  out <- drake_meta_impl(target, config)
-  class(out) <- c("drake_meta", "drake")
-  out
+  spec <- spec %|||% config$spec[[target]]
+  meta <- drake_meta_impl(target, spec, config)
+  meta_old <- NULL
+  if (target_exists(target, config)) {
+    meta_old <- config$cache$get(key = target, namespace = "meta")
+  }
+  meta <- subsume_old_meta(target, meta, meta_old, config)
+  class(meta) <- c("drake_meta", "drake")
+  meta
 }
 
 #' @export
@@ -10,7 +16,7 @@ print.drake_meta <- function(x, ...) {
   cat("drake metadata for ", display_key(x$name), ":\n", sep = "")
   elts <- names(x)
   long <- c("command", "date")
-  lsts <- c("trigger", "time_start", "time_build", "time_command")
+  lsts <- c("trigger", "time_start", "time_build", "time_command", "meta_old")
   list1 <- x[setdiff(elts, c(long, lsts))]
   list2 <- x[intersect(elts, long)]
   list2 <- lapply(list2, crop_text, width = getOption("width") - 18L)
@@ -22,6 +28,12 @@ print.drake_meta <- function(x, ...) {
 
 drake_meta_class <- function(target, config) {
   spec <- config$spec[[target]]
+  if (is_subtarget(target, config)) {
+    return("subtarget")
+  }
+  if (is_dynamic(target, config)) {
+    return("dynamic")
+  }
   if (is_encoded_path(target)) {
     return("imported_file")
   }
@@ -29,21 +41,14 @@ drake_meta_class <- function(target, config) {
   if (is_imported) {
     return("imported_object")
   }
-  if (is_subtarget(target, config)) {
-    return("subtarget")
-  }
-  if (is_dynamic(target, config)) {
-    return("dynamic")
-  }
   "static"
 }
 
-drake_meta_impl <- function(target, config) {
+drake_meta_impl <- function(target, spec, config) {
   UseMethod("drake_meta_impl")
 }
 
-drake_meta_impl.imported_file <- function(target, config) { # nolint
-  spec <- config$spec[[target]]
+drake_meta_impl.imported_file <- function(target, spec, config) { # nolint
   meta <- list(
     name = target,
     target = target,
@@ -57,12 +62,11 @@ drake_meta_impl.imported_file <- function(target, config) { # nolint
   meta$mtime <- storage_mtime(path)
   meta$size_storage <- storage_size(path)
   spec$trigger <- trigger(condition = TRUE)
-  meta <- drake_meta_static_triggers(target, meta, spec, config)
+  meta <- decorate_trigger_meta(target, meta, spec, config)
   meta
 }
 
-drake_meta_impl.imported_object <- function(target, config) { # nolint
-  spec <- config$spec[[target]]
+drake_meta_impl.imported_object <- function(target, spec, config) { # nolint
   meta <- list(
     name = target,
     target = target,
@@ -74,23 +78,25 @@ drake_meta_impl.imported_object <- function(target, config) { # nolint
     file_out = spec$deps_build$file_out
   )
   spec$trigger <- trigger(condition = TRUE)
-  meta <- drake_meta_static_triggers(target, meta, spec, config)
+  meta <- decorate_trigger_meta(target, meta, spec, config)
   meta
 }
 
-drake_meta_impl.subtarget <- function(target, config) {
+drake_meta_impl.subtarget <- function(target, spec, config) {
   list(
     name = target,
     target = target,
     imported = FALSE,
     isfile = FALSE,
+    dynamic = FALSE,
+    format = spec$format %||NA% "none",
     seed = resolve_target_seed(target, config),
-    time_start = drake_meta_start(config)
+    time_start = drake_meta_start(config),
+    trigger = as.list(spec$trigger)
   )
 }
 
-drake_meta_impl.dynamic <- function(target, config) {
-  spec <- config$spec[[target]]
+drake_meta_impl.dynamic <- function(target, spec, config) {
   meta <- list(
     name = target,
     target = target,
@@ -104,11 +110,10 @@ drake_meta_impl.dynamic <- function(target, config) {
     dynamic_dependency_hash = dynamic_dependency_hash(target, config),
     max_expand = spec$max_expand %||NA% config$max_expand
   )
-  drake_meta_static_triggers(target, meta, spec, config)
+  decorate_trigger_meta(target, meta, spec, config)
 }
 
-drake_meta_impl.static <- function(target, config) {
-  spec <- config$spec[[target]]
+drake_meta_impl.static <- function(target, spec, config) {
   meta <- list(
     name = target,
     target = target,
@@ -121,11 +126,11 @@ drake_meta_impl.static <- function(target, config) {
     seed = resolve_target_seed(target, config),
     time_start = drake_meta_start(config)
   )
-  meta <- drake_meta_static_triggers(target, meta, spec, config)
+  meta <- decorate_trigger_meta(target, meta, spec, config)
   meta
 }
 
-drake_meta_static_triggers <- function(target, meta, spec, config) {
+decorate_trigger_meta <- function(target, meta, spec, config) {
   meta$trigger <- as.list(spec$trigger)
   if (meta$trigger$command) {
     meta$command <- spec$command_standardized
@@ -141,6 +146,48 @@ drake_meta_static_triggers <- function(target, meta, spec, config) {
     try_load_deps(spec$deps_change$memory, config = config)
     meta$trigger$value <- eval(meta$trigger$change, config$envir_targets)
   }
+  meta
+}
+
+subsume_old_meta <- function(target, meta, meta_old, config) {
+  meta$meta_old <- meta_old
+  if (!is_dynamic(target, config)) {
+    class(target) <- meta$format
+    meta <- decorate_trigger_format_meta(target, meta, config)
+  }
+  meta
+}
+
+decorate_trigger_format_meta <- function(target, meta, config) {
+  UseMethod("decorate_trigger_format_meta")
+}
+
+decorate_trigger_format_meta.default <- function(target, meta, config) { # nolint
+  meta
+}
+
+decorate_trigger_format_meta.file <- function(target, meta, config) { # nolint
+  if (is.null(meta$meta_old) || !meta$trigger$file) {
+    return(meta)
+  }
+  path <- as.character(meta$meta_old$format_file_path)
+  new_mtime <- storage_mtime(path)
+  new_size <- storage_size(path)
+  hash <- as.character(meta$meta_old$format_file_hash)
+  exists <- file.exists(path)
+  hash[!exists] <- ""
+  should_rehash <- exists & should_rehash_local(
+    size_threshold = rehash_storage_size_threshold,
+    new_mtime = new_mtime,
+    old_mtime = as.numeric(meta$meta_old$format_file_time),
+    new_size = new_size,
+    old_size = as.numeric(meta$meta_old$format_file_size)
+  )
+  hash[should_rehash] <- rehash_local(path[should_rehash], config)
+  meta$format_file_path <- path
+  meta$format_file_hash <- hash
+  meta$format_file_time <- new_mtime
+  meta$format_file_size <- new_size
   meta
 }
 
@@ -171,11 +218,16 @@ target_exists_single <- function(target, config) {
   ht_exists(ht = config$ht_target_exists, x = target)
 }
 
-target_exists_fast <- Vectorize(
+target_exists_fast_list <- Vectorize(
   target_exists_single,
   vectorize.args = "target",
   USE.NAMES = FALSE
 )
+
+target_exists_fast <- function(target, config) {
+  out <- target_exists_fast_list(target, config)
+  as.logical(out)
+}
 
 resolve_target_seed <- function(target, config) {
   seed <- config$spec[[target]]$seed
@@ -322,7 +374,7 @@ static_storage_hash <- function(
   )
 }
 
-should_rehash_local <- function(
+should_rehash_local_impl <- function(
   size_threshold,
   new_mtime,
   old_mtime,
@@ -335,15 +387,44 @@ should_rehash_local <- function(
   small || touched || resized
 }
 
+should_rehash_local_list <- Vectorize(
+  should_rehash_local_impl,
+  vectorize.args = c("new_mtime", "old_mtime", "new_size", "old_size"),
+  USE.NAMES = FALSE
+)
+
+should_rehash_local <- function(
+  size_threshold,
+  new_mtime,
+  old_mtime,
+  new_size,
+  old_size
+) {
+  out <- should_rehash_local_list(
+    size_threshold = size_threshold,
+    new_mtime = new_mtime,
+    old_mtime = old_mtime,
+    new_size = new_size,
+    old_size = old_size
+  )
+  as.logical(out)
+}
+
 rehash_storage_size_threshold <- 1e5
 rehash_storage_size_tol <- .Machine$double.eps ^ 0.5
 
+storage_mtime_impl <- function(x) {
+  ifelse(dir.exists(x), dir_mtime(x), file_mtime(x))
+}
+
+storage_mtime_list <- Vectorize(
+  storage_mtime_impl,
+  vectorize.args = "x",
+  USE.NAMES = FALSE
+)
+
 storage_mtime <- function(x) {
-  if (dir.exists(x)) {
-    dir_mtime(x)
-  } else {
-    file_mtime(x)
-  }
+  as.numeric(storage_mtime_list(x))
 }
 
 dir_mtime <- function(x) {
@@ -362,12 +443,18 @@ file_mtime <- function(x) {
   as.numeric(file.mtime(x))
 }
 
+storage_size_impl <- function(x) {
+  ifelse(dir.exists(x), dir_size(x), file_size(x))
+}
+
+storage_size_list <- Vectorize(
+  storage_size_impl,
+  vectorize.args = "x",
+  USE.NAMES = FALSE
+)
+
 storage_size <- function(x) {
-  if (dir.exists(x)) {
-    dir_size(x)
-  } else {
-    file_size(x)
-  }
+  as.numeric(storage_size_list(x))
 }
 
 dir_size <- function(x) {
@@ -406,12 +493,18 @@ rehash_static_storage <- function(target, file = NULL, config) {
   rehash_local(file, config)
 }
 
+rehash_local_impl <- function(file, config) {
+  ifelse(dir.exists(file), rehash_dir(file, config), rehash_file(file, config))
+}
+
+rehash_local_list <- Vectorize(
+  rehash_local_impl,
+  vectorize.args = "file",
+  USE.NAMES = FALSE
+)
+
 rehash_local <- function(file, config) {
-  if (dir.exists(file)) {
-    rehash_dir(file, config)
-  } else {
-    rehash_file(file, config)
-  }
+  as.character(rehash_local_list(file, config))
 }
 
 rehash_dir <- function(dir, config) {