+++ R/dynamic.R
@@ -213,16 +213,18 @@ chr_dynamic_impl.default <- function(x) {
 
 register_subtargets <- function(target, static_ok, dynamic_ok, config) {
   on.exit(register_dynamic(target, config))
-  announce_build(target, config)
   subtargets_build <- subtargets_all <- subtarget_names(target, config)
   if (static_ok) {
-    subtargets_build <- filter_subtargets(subtargets_all, config)
+    subtargets_build <- filter_subtargets(target, subtargets_all, config)
   }
   if (length(subtargets_all)) {
     register_in_graph(target, subtargets_all, config)
     register_in_spec(target, subtargets_all, config)
   }
   ndeps <- length(subtargets_build)
+  if (ndeps) {
+    announce_build(target, config)
+  }
   if (ndeps) {
     config$logger$disk("register", ndeps, "subtargets", target = target)
     config$logger$inc_progress_total(ndeps)
@@ -239,8 +241,10 @@ register_subtargets <- function(target, static_ok, dynamic_ok, config) {
   }
 }
 
-filter_subtargets <- function(subtargets, config) {
-  subtargets <- subtargets[target_missing(subtargets, config)]
+filter_subtargets <- function(target, subtargets, config) {
+  ht_set(config$ht_is_subtarget, subtargets)
+  index <- check_subtarget_triggers(target, subtargets, config)
+  subtargets <- subtargets[index]
   if (!config$recover || !length(subtargets)) {
     return(subtargets)
   }
@@ -248,7 +252,8 @@ filter_subtargets <- function(subtargets, config) {
     subtargets,
     recover_subtarget,
     jobs = config$jobs_preprocess,
-    config = config
+    config = config,
+    parent = target
   )
   subtargets[!unlist(recovered)]
 }