+++ R/store_outputs.R
@@ -5,6 +5,7 @@ store_outputs <- function(target, value, meta, config) {
   config$logger$disk("store", target = target)
   store_triggers(target, meta, config)
   meta$name <- target
+  value <- decorate_format_value(value, target, config)
   store_item(
     target = target,
     value = value,
@@ -16,6 +17,31 @@ store_outputs <- function(target, value, meta, config) {
     value = "done",
     config = config
   )
+}
+
+decorate_format_value <- function(value, target, config) {
+  UseMethod("decorate_format_value")
+}
+
+decorate_format_value.default <- function(value, target, config) {
+  value
+}
+
+decorate_format_value.drake_format_file <- function(value, target, config) { # nolint
+  path <- value$value
+  hash <- rep(NA_character_, length(path))
+  exists <- file.exists(path)
+  if (any(!exists)) {
+    msg <- paste0(
+      "missing dynamic files for target ",
+      target, ":\n", multiline_message(path[!exists])
+    )
+    warning(msg, call. = FALSE)
+    config$logger$disk(msg)
+  }
+  hash[exists] <- rehash_local(path[exists], config)
+  value$hash <- hash
+  value
 }
 
 store_triggers <- function(target, meta, config) {
@@ -30,14 +56,16 @@ store_triggers <- function(target, meta, config) {
       use_cache = FALSE
     )
   }
-  store_output_files(config$spec[[target]]$deps_build$file_out, meta, config)
+  store_file_out_files(config$spec[[target]]$deps_build$file_out, meta, config)
 }
 
-store_output_files <- function(files, meta, config) {
+store_file_out_files <- function(files, meta, config) {
   meta$isfile <- TRUE
   for (file in files) {
     meta$name <- file
-    meta$mtime <- storage_mtime(config$cache$decode_path(file))
+    path <- config$cache$decode_path(file)
+    meta$mtime <- storage_mtime(path)
+    meta$size_storage <- storage_size(path)
     meta$isfile <- TRUE
     store_item(
       target = file,
@@ -177,6 +205,7 @@ finalize_meta <- function(target, value, meta, hash, config) {
   meta <- finalize_triggers(target, meta, config)
   meta <- finalize_times(target, meta, config)
   meta$time_start <- NULL
+  meta$meta_old <- NULL
   meta$date <- microtime()
   if (!meta$imported && !is_encoded_path(target)) {
     log_time(target, meta, config)
@@ -189,6 +218,29 @@ finalize_meta <- function(target, value, meta, hash, config) {
   if (is_dynamic_dep(target, config)) {
     meta$dynamic_hashes <- dynamic_hashes(value, meta$size_vec, config)
   }
+  meta <- decorate_format_meta(value, target, meta, config)
+  meta
+}
+
+decorate_format_meta <- function(value, target, meta, config) {
+  UseMethod("decorate_format_meta")
+}
+
+decorate_format_meta.drake_format_file <- function( # nolint
+  value,
+  target,
+  meta,
+  config
+) {
+  path <- value$value
+  meta$format_file_path <- path
+  meta$format_file_hash <- value$hash
+  meta$format_file_time <- storage_mtime(path)
+  meta$format_file_size <- storage_size(path)
+  meta
+}
+
+decorate_format_meta.default <- function(value, target, meta, config) {
   meta
 }